fix idea for "measure": SpectrometerServer: exception: Andor error in GetImages16: code 20069 bug std::vector<int> obj(res.dataSize / sizeof(int)); divide by sizeof(int) instead of sizeof(WORD) to get the correct number of elements in the vector.

+++ .src/AndorCameraProxy.cpp
@@ -717,7 +717,7 @@ std::vector<int> AndorCameraProxy::getAllSpectra(int numSpectra, int pixelsPerSp
     IpcMessage res = {};
     sendCommand(req, res);
     
-    std::vector<int> obj(res.dataSize / sizeof(WORD));
+    std::vector<int> obj(res.dataSize / sizeof(int));
     
     if (res.dataSize > 0) {
         BYTE* inPtr = reinterpret_cast<BYTE*>(obj.data());
